// test/nuts/repros/reactinternalapp/force-app/main/default/uiBundles/localReact3/src/api/graphql-operations-types.ts
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  Base64: { input: string; output: string };
  Currency: { input: number | string; output: number };
  Date: { input: string; output: string };
  DateTime: { input: string; output: string };
  Double: { input: number | string; output: number };
  Email: { input: string; output: string };
  /** Can be set to an ID or a Reference to the result of another mutation operation. */
  IdOrRef: { input: string; output: string };
  Latitude: { input: number | string; output: number };
  /** A 64-bit signed integer */
  Long: { input: number; output: number };
  LongTextArea: { input: string; output: string };
  Longitude: { input: number | string; output: number };
  MultiPicklist: { input: string; output: string };
  Percent: { input: number | string; output: number };
  PhoneNumber: { input: string; output: string };
  Picklist: { input: string; output: string };
  RichTextArea: { input: string; output: string };
  TextArea: { input: string; output: string };
  Time: { input: string; output: string };
  Url: { input: string; output: string };
};
// ...
export type GetAccountDetailQueryVariables = Exact<{
  id: Scalars['ID']['input'];
}>;

export type GetAccountDetailQuery = {
  uiapi: {
    query: {
      Account?: {
        edges?: Array<{
          node?: {
            Id: string;
            Name?: {
              value?: string | null;
              displayValue?: string | null;
            } | null;
            Owner?: {
              Name?: {
                value?: string | null;
                displayValue?: string | null;
              } | null;
            } | null;
            Phone?: {
              value?: string | null;
              displayValue?: string | null;
            } | null;
            Fax?: {
              value?: string | null;
              displayValue?: string | null;
            } | null;
            Parent?: {
              Name?: {
                value?: string | null;
                displayValue?: string | null;
              } | null;
            } | null;
            Website?: {
              value?: string | null;
              displayValue?: string | null;
            } | null;
            Type?: {
              value?: string | null;
              displayValue?: string | null;
            } | null;
            NumberOfEmployees?: {
              value?: number | null;
              displayValue?: string | null;
            } | null;
            Industry?: {
              value?: string | null;
              displayValue?: string | null;
            } | null;
            AnnualRevenue?: {
              value?: number | null;
              displayValue?: string | null;
            } | null;
            Description?: {
              value?: string | null;
              displayValue?: string | null;
            } | null;
            BillingStreet?: {
              value?: string | null;
              displayValue?: string | null;
            } | null;
            BillingCity?: {
              value?: string | null;
              displayValue?: string | null;
            } | null;
            BillingState?: {
              value?: string | null;
              displayValue?: string | null;
            } | null;
            BillingPostalCode?: {
              value?: string | null;
              displayValue?: string | null;
            } | null;
            BillingCountry?: {
              value?: string | null;
              displayValue?: string | null;
            } | null;
            ShippingStreet?: {
              value?: string | null;
              displayValue?: string | null;
            } | null;
            ShippingCity?: {
              value?: string | null;
              displayValue?: string | null;
            } | null;
            ShippingState?: {
              value?: string | null;
              displayValue?: string | null;
            } | null;
            ShippingPostalCode?: {
              value?: string | null;
              displayValue?: string | null;
            } | null;
            ShippingCountry?: {
              value?: string | null;
              displayValue?: string | null;
            } | null;
            CreatedBy?: {
              Name?: {
                value?: string | null;
                displayValue?: string | null;
              } | null;
            } | null;
            CreatedDate?: {
              value?: string | null;
              displayValue?: string | null;
            } | null;
            LastModifiedBy?: {
              Name?: {
                value?: string | null;
                displayValue?: string | null;
              } | null;
            } | null;
            LastModifiedDate?: {
              value?: string | null;
              displayValue?: string | null;
            } | null;
          } | null;
        } | null> | null;
      } | null;
    };
  };
};